+++ lib/mongol_text.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:mongol/mongol_font.dart';
 
@@ -8,6 +9,7 @@ class MongolText extends StatelessWidget {
     this.data, {
     Key key,
     this.style,
+    this.textScaleFactor,
   })  : assert(data != null, 'The data cannot be null.'),
         super(key: key);
 
@@ -17,6 +19,9 @@ class MongolText extends StatelessWidget {
   /// This is the style to use for the whole text string. If null a default
   /// style will be used.
   final TextStyle style;
+
+  /// Font pixels per logical pixel
+  final double textScaleFactor;
 
   static const TextStyle _defaultMongolTextStyle =
       TextStyle(fontFamily: MongolFont.qagan, fontSize: 24.0);
@@ -35,7 +40,17 @@ class MongolText extends StatelessWidget {
         text: data,
         style: effectiveTextStyle,
       ),
+      textScaleFactor: textScaleFactor ?? MediaQuery.textScaleFactorOf(context),
     );
     return result;
+  }
+
+  @override
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties.add(StringProperty('data', data, showName: false));
+    style?.debugFillProperties(properties);
+    properties.add(
+        DoubleProperty('textScaleFactor', textScaleFactor, defaultValue: 1.0));
   }
 }